sensornet-decode-server/requirements.txt
certifi==2022.12.7
influxdb-client==1.35.0
numpy==1.23.5
pandas==1.5.2
python-dateutil==2.8.2
python-dotenv==0.21.0
pytz==2022.6
reactivex==4.0.4
six==1.16.0
typing_extensions==4.4.0
urllib3==1.26.13
